+++ core/api/typescript/src/lib/enclaves/grpc_node_api_container_client.ts
@@ -1,4 +1,4 @@
-import { ok, err, Result } from "neverthrow";
+import {ok, err, Result, Err} from "neverthrow";
 import type { ServiceError } from "@grpc/grpc-js";
 import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
 import {
@@ -23,8 +23,12 @@ import {
     ExecuteModuleArgs,
     ExecCommandArgs,
     ExecCommandResponse,
+    UploadFilesArtifactArgs,
+    UploadFilesArtifactResponse,
     StoreWebFilesArtifactArgs,
-    StoreWebFilesArtifactResponse, StoreFilesArtifactFromServiceArgs, StoreFilesArtifactFromServiceResponse,
+    StoreWebFilesArtifactResponse,
+    StoreFilesArtifactFromServiceArgs,
+    StoreFilesArtifactFromServiceResponse,
 } from "../../kurtosis_core_rpc_api_bindings/api_container_service_pb";
 import type { ApiContainerServiceClient as ApiContainerServiceClientNode } from "../../kurtosis_core_rpc_api_bindings/api_container_service_grpc_pb";
 import { GenericApiContainerClient } from "./generic_api_container_client";
@@ -366,6 +370,29 @@ export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
         return ok(execCommandResponse)
     }
 
+    public async uploadFiles(uploadFilesArtifactArgs: UploadFilesArtifactArgs): Promise<Result<UploadFilesArtifactResponse, Error>> {
+        const uploadFilesArtifactPromise: Promise<Result<UploadFilesArtifactResponse, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.uploadFilesArtifact(uploadFilesArtifactArgs, (error: ServiceError | null, response?: UploadFilesArtifactResponse) => {
+                if (error === null) {
+                    if (!response) {
+                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
+                    } else {
+                        resolve(ok(response!));
+                    }
+                } else {
+                    resolve(err(error));
+                }
+            })
+        });
+        const uploadFilesArtifactResponseResult: Result<UploadFilesArtifactResponse, Error> = await uploadFilesArtifactPromise;
+        if(uploadFilesArtifactResponseResult.isErr()){
+            return err(uploadFilesArtifactResponseResult.error)
+        }
+
+        const uploadFilesArtifactResponse = uploadFilesArtifactResponseResult.value
+        return ok(uploadFilesArtifactResponse)
+    }
+
     public async storeWebFilesArtifact(storeWebFilesArtifactArgs: StoreWebFilesArtifactArgs): Promise<Result<StoreWebFilesArtifactResponse, Error>> {
         const storeWebFilesArtifactPromise: Promise<Result<StoreWebFilesArtifactResponse, Error>> = new Promise((resolve, _unusedReject) => {
             this.client.storeWebFilesArtifact(storeWebFilesArtifactArgs, (error: ServiceError | null, response?: StoreWebFilesArtifactResponse) => {
@@ -380,6 +407,7 @@ export class GrpcNodeApiContainerClient implements GenericApiContainerClient {
                 }
             })
         });
+
         const storeWebFilesArtifactResponseResult: Result<StoreWebFilesArtifactResponse, Error> = await storeWebFilesArtifactPromise;
         if(storeWebFilesArtifactResponseResult.isErr()){
             return err(storeWebFilesArtifactResponseResult.error)