+++ core/api/typescript/src/lib/enclaves/grpc_web_api_container_client.ts
@@ -1,4 +1,4 @@
-import { ok, err, Result } from "neverthrow";
+import {ok, err, Result, Err} from "neverthrow";
 import * as grpc_web from "grpc-web";
 import * as google_protobuf_empty_pb from "google-protobuf/google/protobuf/empty_pb";
 import {
@@ -23,6 +23,8 @@ import {
     ExecuteModuleResponse,
     ExecCommandArgs,
     ExecCommandResponse,
+    UploadFilesArtifactArgs,
+    UploadFilesArtifactResponse,
     StoreWebFilesArtifactResponse,
     StoreWebFilesArtifactArgs,
     StoreFilesArtifactFromServiceArgs,
@@ -368,6 +370,29 @@ export class GrpcWebApiContainerClient implements GenericApiContainerClient {
         return ok(execCommandResponse)
     }
 
+    public async uploadFiles(uploadFilesArtifactArgs: UploadFilesArtifactArgs): Promise<Result<UploadFilesArtifactResponse, Error>> {
+            const uploadFilesArtifactPromise: Promise<Result<UploadFilesArtifactResponse, Error>> = new Promise((resolve, _unusedReject) => {
+            this.client.uploadFilesArtifact(uploadFilesArtifactArgs, {}, (error: grpc_web.RpcError | null, response?: UploadFilesArtifactResponse) => {
+                if (error === null) {
+                    if (!response) {
+                        resolve(err(new Error("No error was encountered but the response was still falsy; this should never happen")));
+                    } else {
+                        resolve(ok(response!));
+                    }
+                } else {
+                    resolve(err(error));
+                }
+            })
+        });
+        const uploadFilesArtifactResponseResult = await uploadFilesArtifactPromise;
+        if(uploadFilesArtifactResponseResult.isErr()){
+            return err(uploadFilesArtifactResponseResult.error)
+        }
+
+        const uploadFilesArtifactResponse = uploadFilesArtifactResponseResult.value
+        return ok(uploadFilesArtifactResponse)
+    }
+  
     public async storeWebFilesArtifact(storeWebFilesArtifactArgs: StoreWebFilesArtifactArgs): Promise<Result<StoreWebFilesArtifactResponse, Error>> {
         const storeWebFilesArtifactPromise: Promise<Result<StoreWebFilesArtifactResponse, Error>> = new Promise((resolve, _unusedReject) => {
             this.client.storeWebFilesArtifact(storeWebFilesArtifactArgs, {}, (error: grpc_web.RpcError | null, response?: StoreWebFilesArtifactResponse) => {